pyproject.toml
[project]
name = "abelohost"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
linter = [
    "black>=26.3.1",
    "pre-commit>=4.5.1",
    "pylint>=4.0.5",
    "ruff>=0.15.9",
]
main = [
    "alembic>=1.18.4",
    "asyncio>=4.0.0",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.3",
    "granian>=2.7.2",
    "itsdangerous>=2.2.0",
    "orjson>=3.11.8",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic-settings>=2.13.1",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "structlog>=25.5.0",
    "uuid7>=0.1.0",
]
test = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]
[tool.setuptools]
packages = ["backend"]
[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  | alembic
  | env\.py
  | scripts
)/
'''
[tool.ruff]
target-version = "py313"
line-length = 88
src = ["backend"]
exclude = [
    "alembic",
    "env.py",
    "scripts"
]
# Включаем нужные правила
[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "TID", # flake8-tidy-imports
    "UP",  # pyupgrade
    "RUF", # Ruff-specific
    "SIM", # simplify
    "PTH", # pathlib
    "ARG", # unused arguments (лучше, чем pylint для async)
]
ignore = [
    "E501",   # line too long — Black сам управляет длиной
    "B008",   # не вызывать сложные объекты в аргументах по умолчанию (иногда нужно для FastAPI)
    "UP037",  # автоматически удаляет кавычки вокруг типов в аннотациях, если считает, что тип уже доступен в области видимости
    "TID252", # требует только абсолютные пути
]

[tool.ruff.lint.isort]
known-first-party = ["abelohost"]
combine-as-imports = true
split-on-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
# Например, игнорировать "unused args" в роутах или тестах
"tests/**.py" = ["ARG001", "ARG002"]

[tool.ruff.lint.pydocstyle]
convention = "google"  # или "numpy", "pep257" — по желанию
[tool.pylint.'MESSAGES CONTROL']
disable = [
    "C0114", # missing-module-docstring
    "C0115", # missing-class-docstring
    "C0116", # missing-function-docstring
    "C0301", # line-too-long (Black управляет этим)
    "R0903", # too-few-public-methods (часто у dataclasses/Pydantic)
    "R0913", # too-many-arguments (иногда неизбежно в FastAPI)
    "W0511", # fixme (можно оставить, если хотите)
]

[tool.pylint.'DESIGN']
max-args = 8
max-locals = 20
max-returns = 6

[tool.pylint.'FORMAT']
max-line-length = 88

[tool.pylint.'MASTER']
init-hook = 'import sys; sys.path.append(".")'
extension-pkg-allow-list = ["pydantic", "orjson"]
ignore-paths = [
  ".*/tests/.*",
  ".*/test_.*\\.py",
  ".*/.*_test\\.py"
]
